+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/BacklinksDelegateImpl.kt
@@ -21,36 +21,63 @@ import io.realm.kotlin.internal.schema.ClassMetadata
 import io.realm.kotlin.internal.schema.PropertyMetadata
 import io.realm.kotlin.query.RealmResults
 import io.realm.kotlin.types.BacklinksDelegate
+import io.realm.kotlin.types.EmbeddedBacklinksDelegate
+import io.realm.kotlin.types.EmbeddedRealmObject
 import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.TypedRealmObject
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty
 
-internal class BacklinksDelegateImpl<T : TypedRealmObject>(private val sourceClass: KClass<T>) : BacklinksDelegate<T> {
-    override fun getValue(
-        reference: RealmObject,
+internal class BacklinksDelegateImpl<T : TypedRealmObject>(
+    private val sourceClass: KClass<T>
+) : EmbeddedBacklinksDelegate<T>, BacklinksDelegate<T> {
+    private fun createBacklinks(
+        reference: TypedRealmObject,
         targetProperty: KProperty<*>
     ): RealmResults<T> {
         if (!reference.isManaged()) {
             throw IllegalStateException("Unmanaged objects don't support backlinks.")
         }
         val objectReference = reference.realmObjectReference!!
-        val targetPropertyMetadata: PropertyMetadata =
-            objectReference.metadata[targetProperty]!!
+
+        // Target property must be part of the reference object.
+        val targetPropertyMetadata: PropertyMetadata = objectReference.metadata[targetProperty]
+            ?: throw IllegalArgumentException("Target property '${targetProperty.name}' not defined in '${reference::class.simpleName}'.")
+
+        // Target property must be a backlink.
+        targetPropertyMetadata.linkOriginPropertyName.ifEmpty {
+            throw IllegalArgumentException("Target property '${targetProperty.name}' is not a backlink property.")
+        }
 
         val sourceClassMetadata: ClassMetadata = objectReference.owner
             .schemaMetadata
             .getOrThrow(targetPropertyMetadata.linkTarget)
 
-        val sourcePropertyKey = sourceClassMetadata[targetPropertyMetadata.linkOriginPropertyName]!!.key
+        // Delegate type parameter T and source class must match.
+        if (sourceClass != sourceClassMetadata.clazz) {
+            throw IllegalArgumentException("Target property type '${sourceClassMetadata.clazz!!.simpleName}' does not match backlink type '${sourceClass.simpleName}'.")
+        }
+
+        val sourcePropertyKey = sourceClassMetadata[targetPropertyMetadata.linkOriginPropertyName]!!
 
         val linkingObjects: RealmResultsImpl<T> = RealmObjectHelper.getBacklinks(
             obj = objectReference,
-            sourcePropertyKey = sourcePropertyKey,
+            sourcePropertyKey = sourcePropertyKey.key,
             sourceClassKey = sourceClassMetadata.classKey,
             sourceClass = sourceClass
         )
 
         return ObjectBoundRealmResults(objectReference, linkingObjects)
     }
+
+    override fun getValue(
+        reference: EmbeddedRealmObject,
+        targetProperty: KProperty<*>
+    ): T = createBacklinks(reference, targetProperty).firstOrNull()
+        ?: throw IllegalStateException("Backlink '${targetProperty.name}' is not an instance of target property type '${sourceClass!!.simpleName}'.")
+
+    override fun getValue(
+        reference: RealmObject,
+        targetProperty: KProperty<*>
+    ): RealmResults<T> = createBacklinks(reference, targetProperty)
 }