+++ packages/test-base/src/androidAndroidTest/kotlin/io/realm/kotlin/test/shared/RealmListTests.kt
@@ -27,6 +27,7 @@ import io.realm.kotlin.entities.list.Level2
 import io.realm.kotlin.entities.list.Level3
 import io.realm.kotlin.entities.list.RealmListContainer
 import io.realm.kotlin.entities.list.listTestSchema
+import io.realm.kotlin.ext.asRealmObject
 import io.realm.kotlin.ext.query
 import io.realm.kotlin.ext.realmListOf
 import io.realm.kotlin.ext.toRealmList
@@ -37,6 +38,7 @@ import io.realm.kotlin.test.assertFailsWithMessage
 import io.realm.kotlin.test.platform.PlatformUtils
 import io.realm.kotlin.test.util.TypeDescriptor
 import io.realm.kotlin.types.ObjectId
+import io.realm.kotlin.types.RealmAny
 import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
@@ -64,6 +66,7 @@ import kotlin.test.assertFailsWith
 import kotlin.test.assertFalse
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
+import kotlin.test.fail
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
 
@@ -636,6 +639,7 @@ class RealmListTests {
         RealmUUID::class -> if (nullable) NULLABLE_UUID_VALUES else UUID_VALUES
         ByteArray::class -> if (nullable) NULLABLE_BINARY_VALUES else BINARY_VALUES
         RealmObject::class -> OBJECT_VALUES
+        RealmAny::class -> LIST_REALM_ANY_VALUES
         else -> throw IllegalArgumentException("Wrong classifier: '$classifier'")
     } as List<T>
 
@@ -671,6 +675,14 @@ class RealmListTests {
                         classifier,
                         elementType.nullable
                     ) as TypeSafetyManager<ByteArray?>
+                )
+                RealmAny::class -> ManagedRealmAnyListTester(
+                    realm = realm,
+                    typeSafetyManager = NullableList(
+                        classifier = classifier,
+                        property = RealmListContainer.nullableProperties[classifier]!!,
+                        dataSet = getDataSetForClassifier(classifier, true)
+                    ) as TypeSafetyManager<RealmAny?>
                 )
                 else -> ManagedGenericListTester(
                     realm = realm,
@@ -1247,7 +1259,7 @@ internal abstract class ManagedListTester<T>(
             }
 
             // Clean up
-            delete(findLatest(container)!!)
+            delete(query<RealmListContainer>())
         }
     }
 
@@ -1264,7 +1276,7 @@ internal abstract class ManagedListTester<T>(
 
         // Clean up
         realm.writeBlocking {
-            delete(findLatest(container)!!)
+            delete(query<RealmListContainer>())
         }
     }
 }
@@ -1273,7 +1285,7 @@ internal abstract class ManagedListTester<T>(
  * No special needs for managed, generic testers. Elements can be compared painlessly and need not
  * be copied to Realm when calling RealmList API methods.
  */
-internal class ManagedGenericListTester<T>(
+internal class ManagedGenericListTester<T> constructor(
     realm: Realm,
     typeSafetyManager: TypeSafetyManager<T>
 ) : ManagedListTester<T>(realm, typeSafetyManager) {
@@ -1282,6 +1294,43 @@ internal class ManagedGenericListTester<T>(
             assertContentEquals(expected, actual as ByteArray)
         } else {
             assertEquals(expected, actual)
+        }
+    }
+}
+
+/**
+ * Checks equality for RealmAny values. When working with RealmObjects we need to do it at a
+ * structural level.
+ */
+internal class ManagedRealmAnyListTester constructor(
+    realm: Realm,
+    typeSafetyManager: TypeSafetyManager<RealmAny?>
+) : ManagedListTester<RealmAny?>(realm, typeSafetyManager) {
+    override fun assertElementsAreEqual(expected: RealmAny?, actual: RealmAny?) {
+        if (expected != null && actual != null) {
+            assertEquals(expected.type, actual.type)
+            when (expected.type) {
+                RealmAny.Type.INT -> assertEquals(expected.asInt(), actual.asInt())
+                RealmAny.Type.BOOL -> assertEquals(expected.asBoolean(), actual.asBoolean())
+                RealmAny.Type.STRING -> assertEquals(expected.asString(), actual.asString())
+                RealmAny.Type.BINARY ->
+                    assertContentEquals(expected.asByteArray(), actual.asByteArray())
+                RealmAny.Type.TIMESTAMP ->
+                    assertEquals(expected.asRealmInstant(), actual.asRealmInstant())
+                RealmAny.Type.FLOAT -> assertEquals(expected.asFloat(), actual.asFloat())
+                RealmAny.Type.DOUBLE -> assertEquals(expected.asDouble(), actual.asDouble())
+                RealmAny.Type.OBJECT_ID -> assertEquals(expected.asObjectId(), actual.asObjectId())
+                RealmAny.Type.UUID -> assertEquals(
+                    expected.asRealmUUID(),
+                    actual.asRealmUUID()
+                )
+                RealmAny.Type.OBJECT -> assertEquals(
+                    expected.asRealmObject<RealmListContainer>().stringField,
+                    actual.asRealmObject<RealmListContainer>().stringField
+                )
+            }
+        } else if (expected != null || actual != null) {
+            fail("One of the RealmAny values is null, expected = $expected, actual = $actual")
         }
     }
 }
@@ -1346,6 +1395,34 @@ internal val OBJECT_VALUES3 = listOf(
     RealmListContainer().apply { stringField = "H" }
 )
 internal val BINARY_VALUES = listOf(Random.Default.nextBytes(2), Random.Default.nextBytes(2))
+
+// Base RealmAny values. The list does not include 'RealmAny.create(RealmObject())' since it is used
+// as a base for both lists and sets and they use different container classes in their logic.
+// Do NOT use this list directly in your tests unless you have a good reason to ignore RealmAny
+// instances containing a RealmObject.
+internal val REALM_ANY_PRIMITIVE_VALUES = listOf(
+    RealmAny.create((-12).toShort()),
+    RealmAny.create(13),
+    RealmAny.create(14.toByte()),
+    RealmAny.create(15.toChar()),
+    RealmAny.create(16L),
+    RealmAny.create(false),
+    RealmAny.create("Hello"),
+    RealmAny.create(17F),
+    RealmAny.create(18.0),
+    RealmAny.create(BsonObjectId("507f191e810c19729de860ea")),
+    RealmAny.create(byteArrayOf(19)),
+    RealmAny.create(RealmInstant.from(42, 420)),
+    RealmAny.create(RealmUUID.from("46423f1b-ce3e-4a7e-812f-004cf9c42d76")),
+    null
+)
+internal val REALM_ANY_REALM_OBJECT = RealmAny.create(
+    RealmListContainer().apply { stringField = "hello" },
+    RealmListContainer::class
+)
+
+// Use this for LIST tests as this file does exhaustive testing on all RealmAny types
+internal val LIST_REALM_ANY_VALUES = REALM_ANY_PRIMITIVE_VALUES + REALM_ANY_REALM_OBJECT
 
 internal val NULLABLE_CHAR_VALUES = CHAR_VALUES + null
 internal val NULLABLE_STRING_VALUES = STRING_VALUES + null